Refactor asset cards into modular controller
