+++ src/ui/views/classic/assetCards/index.js
@@ -1,0 +1,16 @@
+export { default } from './controller.js';
+export {
+  renderAssets,
+  updateAssets,
+  updateAssetGroup,
+  updateAssetHub,
+  updateAssetEmptyNotice,
+  openInstanceDetails,
+  openAssetGroupDetails,
+  isAssetDefinition
+} from './controller.js';
+
+export * from './cache.js';
+export * from './qualityActions.js';
+export * from './nicheAssignments.js';
+export * from './renderers.js';
